fix(types): remove types-folders and its contents
