+++ src/components/navigation/BreadcrumbEllipsis/stories/BreadcrumbEllipsis.Default.stories.tsx
@@ -3,12 +3,12 @@ import { IBreadcrumbEllipsisProps } from "../interfaces/BreadcrumbEllipsis.inter
 
 import { BreadcrumbEllipsis } from "../index";
 
-import { parameters, routes, size } from "./props";
+import { props } from "../props";
 
 const story = {
   title: "navigation/BreadcrumbEllipsis",
   components: [BreadcrumbEllipsis],
-  parameters,
+  argTypes: props,
   decorators: [
     (Story: React.ElementType) => (
       <div style={{ margin: "2em 2em 7em 2em" }}>
@@ -38,10 +38,6 @@ Default.args = {
     },
   ],
   size: "large",
-};
-Default.argTypes = {
-  routes,
-  size,
 };
 
 export default story;
