--- src/components/navigation/BreadcrumbEllipsis/types/BreadcrumbEllipsis.Typos.type.ts
@@ -1,2 +1,0 @@
-export const sizes = ["large", "small"] as const;
-export type Typos = typeof sizes[number];
