+++ src/components/navigation/BreadcrumbLink/props.ts
@@ -1,0 +1,46 @@
+export const typos = ["large", "medium", "small"] as const;
+export type Typos = typeof typos[number];
+
+const props = {
+  parameters: {
+    docs: {
+      description: {
+        component:
+          "Breadcrumbs are a navigation system used to show a user's location in a site or app.",
+      },
+    },
+  },
+  id: {
+    control: { type: "text" },
+    description: "shall be the id for the text",
+  },
+  isActive: {
+    options: [true, false],
+    control: { type: "boolean" },
+    description:
+      "if the switch is disabled or not. This prevents any interaction.",
+    table: {
+      defaultValue: { summary: "false" },
+    },
+  },
+  handleClick: {
+    options: ["logState"],
+    control: { type: "func" },
+    description: "shall be determine the behavior of the click event",
+  },
+  label: {
+    control: { type: "text" },
+    description: "shall constitute the content to be displayed",
+  },
+  typo: {
+    options: typos,
+    control: { type: "select" },
+    description: "indicates the font size used in the component",
+  },
+  path: {
+    control: { type: "text" },
+    description: "is the path where the BreadcrumbLink is going to navigate",
+  },
+};
+
+export { props };
