--- src/components/navigation/BreadcrumbLink/types/BreadcrumbLink.typos.type.ts
@@ -1,2 +1,0 @@
-export const typos = ["large", "medium", "small"] as const;
-export type Typos = typeof typos[number];
