--- src/components/navigation/BreadcrumbMenu/stories/props.ts
@@ -1,23 +1,0 @@
-const parameters = {
-  docs: {
-    description: {
-      component:
-        "Breadcrumbs are a navigation system used to show a user's location in a site or app.",
-    },
-  },
-};
-
-const routes = {
-  description:
-    "The breadcrumb-component will utilize this routes for display, and consequently, it can be employed to locate the source of breadcrumbLinks present within this component.",
-  table: {
-    defaultValue: {
-      summary: `This structure must be present [{
-         label: "",
-        path: "",
-        id: ""}]`,
-    },
-  },
-};
-
-export { parameters, routes };
