+++ src/components/navigation/Breadcrumbs/index.tsx
@@ -1,11 +1,11 @@
 import { IBreadcrumbsProps } from "./interfaces/Breadcrumbs.interface";
-import { Sizes } from "./types/Breadcrumb.Size.type";
 
 import { BreadcrumbLink } from "../BreadcrumbLink";
 import { BreadcrumbEllipsis } from "../BreadcrumbEllipsis";
 import { useMediaQuery } from "../../../hooks/useMediaQuery";
 
 import { StyledBreadcrumbs } from "./styles";
+import { Sizes } from "./props";
 
 function capitalizeString(string: string) {
   return string.charAt(0).toUpperCase() + string.slice(1);
