+++ src/components/navigation/Breadcrumbs/stories/Breadcrumbs.Default.stories.tsx
@@ -1,12 +1,12 @@
 import { Breadcrumbs } from "../index";
 import { BrowserRouter } from "react-router-dom";
-import { parameters, crumbs } from "./props";
+import { props } from "../props";
 import { IBreadcrumbsProps } from "../interfaces/Breadcrumbs.interface";
 
 const story = {
   title: "navigation/Breadcrumbs",
   components: [Breadcrumbs],
-  parameters,
+  argTypes: props,
   decorators: [
     (Story: React.ElementType) => (
       <div style={{ margin: "3em" }}>
@@ -40,9 +40,6 @@ Default.args = {
       isActive: true,
     },
   ],
-};
-Default.argTypes = {
-  crumbs,
 };
 
 export default story;
