--- src/components/navigation/Breadcrumbs/stories/props.ts
@@ -1,15 +1,0 @@
-const parameters = {
-  docs: {
-    description: {
-      component:
-        "Breadcrumbs are a navigation system used to show a user's location in a site or app.",
-    },
-  },
-};
-
-const crumbs = {
-  description:
-    "An array of objects that contain the path, label, id, and isActive properties.",
-};
-
-export { parameters, crumbs };
