--- src/components/navigation/Breadcrumbs/types/Breadcrumb.Size.type.ts
@@ -1,2 +1,0 @@
-export const sizes = ["large", "small"] as const;
-export type Sizes = typeof sizes[number];
