+++ src/components/navigation/FullscreenNav/props.ts
@@ -1,0 +1,22 @@
+const props = {
+  parameters: {
+    docs: {
+      description: {
+        component:
+          "A versatile side navigation component, which allows you to nest links, from a full screen perspective",
+      },
+    },
+  },
+  navigation: {
+    description:
+      "(objects): shall be designed to accept an array of objects with a predetermined structure, as specified below: Each object shall contain the following attributes",
+  },
+  portalId: {
+    description: "id of the portal element",
+  },
+  logoutPath: {
+    description: "path to the logout page",
+  },
+};
+
+export { props };
