+++ src/components/navigation/FullscreenNav/stories/FullscreenNav.WithoutSections.stories.tsx
@@ -11,13 +11,13 @@ import {
 } from "react-icons/md";
 
 import { FullscreenNav } from "..";
-import { parameters, portalId, navigation, logoutPath } from "./props";
+import { props } from "../props";
 import { IFullscreenNavProps } from "../interfaces/FullscreenNav.interface";
 
 const story = {
   title: "navigation/FullscreenNav",
   components: [FullscreenNav],
-  parameters,
+  argTypes: props,
   decorators: [
     (Story: React.ElementType) => (
       <div style={{ margin: "3em" }}>
@@ -89,12 +89,6 @@ WithoutSections.args = {
   },
   logoutPath: "/logout",
   logoutTitle: "logout",
-};
-
-WithoutSections.argTypes = {
-  portalId,
-  navigation,
-  logoutPath,
 };
 
 export default story;
