--- src/components/navigation/FullscreenNav/stories/props.ts
@@ -1,23 +1,0 @@
-const parameters = {
-  docs: {
-    description: {
-      component:
-        "A versatile side navigation component, which allows you to nest links, from a full screen perspective",
-    },
-  },
-};
-
-const navigation = {
-  description:
-    "(objects): shall be designed to accept an array of objects with a predetermined structure, as specified below: Each object shall contain the following attributes",
-};
-
-const portalId = {
-  description: "id of the portal element",
-};
-
-const logoutPath = {
-  description: "path to the logout page",
-};
-
-export { parameters, portalId, navigation, logoutPath };
