+++ src/components/navigation/Tabs/stories/TabsResponsiveController.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 import { Tabs } from "../index";
-import { Types, types } from "../types/Tabs.type";
+import { Types, types } from "../props";
 import { ITabsProps } from "../interfaces/Tabs.interface";
 
 const TabsResponsiveController = (props: ITabsProps) => {
