--- src/components/navigation/Tabs/types/Tabs.type.ts
@@ -1,3 +1,0 @@
-export const types = ["select", "tabs"] as const;
-
-export type Types = typeof types[number];
