+++ src/hooks/stories/Example.stories.tsx
@@ -4,10 +4,12 @@ import { Switch } from "../../components/inputs/Switch";
 import { useMediaQuery } from "../useMediaQuery";
 import { IExampleSwitch } from "./IExampleSwitch.interface";
 import { SwitchController } from "@src/components/inputs/Switch/stories/SwitchController";
+import { props } from "@src/components/inputs/Switch/props";
 
 const story = {
   title: "hooks/useMediaQuery",
   components: [Switch],
+  argTypes: props,
   decorators: [
     (Story: React.ElementType) => (
       <div style={{ margin: "3em" }}>
@@ -34,9 +36,15 @@ export const Example = (args: IExampleSwitch) => {
 };
 
 Example.args = {
+  id: "id",
+  isDisabled: false,
+  name: "name",
+  value: "switchTest4",
+  checked: false,
+  size: "small",
   handleChange: () => {},
-  id: "thisIsAnId",
-  label: "",
+  margin: "0px",
+  padding: "0px",
 };
 
 export default story;
