--- src/hooks/types/MediaQueryList.type.ts
@@ -1,3 +1,0 @@
-import { IMediaQueryObject } from "../interfaces/MediaQueryObject.interface";
-
-export type MediaQueryList = Array<IMediaQueryObject>;
